refactor: move listing page as first architecture level
