+++ packages/manager/modules/vps/src/vps/translations/Messages_fr_CA.json
