+++ packages/manager/modules/vps/src/vps/vps.less
@@ -1,8 +1,8 @@
 .cloud-vps-section {
   @import 'font-awesome/less/font-awesome.less';
   @import 'flag-icon-css/less/flag-icon.less';
-  @import 'dashboard/vpsTileStatus/vps-tile-status.less';
-  @import 'upgrade/vps-upgrade.less';
+  @import '../dashboard/vpsTileStatus/vps-tile-status.less';
+  @import '../upgrade/vps-upgrade.less';
 
   .vps-dl-horizontal {
     display: flex;
